python-selenium/requirements.txt
async-generator==1.10    
attrs==23.1.0  
bcrypt==4.0.1   
beautifulsoup4==4.12.2  
blinker==1.6.2   
bs4==0.0.1   
certifi==2023.5.7
charset-normalizer==3.1.0   
click==8.1.3   
dnspython==2.3.0   
exceptiongroup==1.1.1   
Flask==2.3.2   
Flask-Cors==3.0.10  
Flask-PyMongo==2.3.0   
h11==0.14.0  
idna==3.4     
importlib-metadata==6.6.0   
itsdangerous==2.1.2   
Jinja2==3.1.2   
MarkupSafe==2.1.2   
outcome==1.2.0   
packaging==23.1    
pip==20.0.2  
pkg-resources==0.0.0   
PyJWT==2.7.0   
pymongo==4.3.3   
PySocks==1.7.1   
python-dotenv==1.0.0   
requests==2.31.0  
selenium==4.9.1   
setuptools==44.0.0  
six==1.16.0  
sniffio==1.3.0   
sortedcontainers==2.4.0   
soupsieve==2.4.1   
tqdm==4.65.0  
trio==0.22.0  
trio-websocket==0.10.2  
urllib3==2.0.2   
webdriver-manager==3.8.6   
Werkzeug==2.3.4   
wsproto==1.2.0   
zipp==3.15.0  
